ci: use immutable fixture and satisfy Linux Clippy

+++ src/job/execution_domain/workspace_reader.rs
@@ -297,17 +297,14 @@ fn open_root(path: &Path) -> Result<RootBinding, ExecutionDomainError> {
     }
     Ok(RootBinding {
         fd,
-        device: metadata.st_dev as u64,
+        device: metadata.st_dev,
         inode: metadata.st_ino as u64,
     })
 }
 
 fn verify_directory(fd: &OwnedFd, device: u64, inode: u64) -> Result<(), ExecutionDomainError> {
     let current = metadata(fd.as_raw_fd())?;
-    if file_type(&current) != libc::S_IFDIR
-        || current.st_dev as u64 != device
-        || current.st_ino != inode
-    {
+    if file_type(&current) != libc::S_IFDIR || current.st_dev != device || current.st_ino != inode {
         return Err(failure(FailureCategory::IdentityMismatch));
     }
     Ok(())
@@ -341,7 +338,7 @@ impl Traversal<'_> {
             check_deadline(self.started, self.limits.max_time)?;
             let before = metadata_at(directory, name)?;
             check_deadline(self.started, self.limits.max_time)?;
-            if before.st_dev as u64 != self.root_device {
+            if before.st_dev != self.root_device {
                 return Err(failure(FailureCategory::IdentityMismatch));
             }
             let relative = relative_dir.join(OsString::from_vec(name.to_bytes().to_vec()));
@@ -359,7 +356,7 @@ impl Traversal<'_> {
                         self.limits.max_time,
                     )?;
                     let after = metadata(child.as_raw_fd())?;
-                    if !same_identity(&before, &after) || after.st_dev as u64 != self.root_device {
+                    if !same_identity(&before, &after) || after.st_dev != self.root_device {
                         return Err(failure(FailureCategory::IdentityMismatch));
                     }
                     self.enumerate(child.as_raw_fd(), &relative, child_depth)?;
@@ -440,7 +437,7 @@ fn open_relative_regular(
             max_time,
         )?;
         let metadata = metadata(next.as_raw_fd())?;
-        if metadata.st_dev as u64 != root.device
+        if metadata.st_dev != root.device
             || if last {
                 file_type(&metadata) != libc::S_IFREG
             } else {
